Implemented more batchbuffer commands in parser, fixed graphics stack build script

+++ common/utils/utils.h
@@ -4,6 +4,6 @@
 #include <inttypes.h>
 
 void find_elf_magic_bytes(pid_t pid, char debug);
-void dump_buffer(unsigned char *kernel, uint64_t size, uint32_t id);
+void dump_buffer(unsigned char *kernel, uint64_t size, uint64_t id);
 void print_map(pid_t pid);
 unsigned char *copy_buffer(uint32_t pid, uint64_t ptr, uint64_t size, char debug);
